killing too-long files; cleaning up APIs
