+++ packages/user/Chat/service/src/tests/mod.rs
@@ -1,0 +1,8 @@
+#![allow(non_snake_case)]
+
+mod helpers;
+
+mod av_call;
+mod chat_data;
+mod group;
+mod spaces;
